+++ app/src/main/java/com/hbb20/androidcountrypicker/CustomRecyclerViewActivity.kt
@@ -3,7 +3,7 @@ package com.hbb20.androidcountrypicker
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.hbb20.countrypicker.loadCountries
+import com.hbb20.countrypicker.recyclerview.loadCountries
 import kotlinx.android.synthetic.main.activity_custom_recycler_view.*
 
 class CustomRecyclerViewActivity : AppCompatActivity() {
