+++ app/src/main/java/com/hbb20/androidcountrypicker/OpenDialogDirectlyActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.hbb20.countrypicker.launchCountryPickerDialog
+import com.hbb20.countrypicker.dialog.launchCountryPickerDialog
 
 class OpenDialogDirectlyActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
