+++ countrypicker/src/main/java/com/hbb20/countrypicker/CPFlagProvider.kt
@@ -1,13 +1,13 @@
 package com.hbb20.countrypicker
 
-sealed class FlagProvider
+sealed class CPFlagProvider
 /**
  * If you wish to useEmojiCompat = true then make sure to enable emoji-compat in your app using
  * [Downloadable font configuration](https://developer.android.com/guide/topics/ui/look-and-feel/emoji-compat#downloadable-fonts)
  * or [Bundled Font Configuration](https://developer.android.com/guide/topics/ui/look-and-feel/emoji-compat#bundled-fonts)
  */
-class DefaultEmojiFlagProvider(val useEmojiCompat: Boolean = false) : FlagProvider()
+class DefaultEmojiFlagProvider(val useEmojiCompat: Boolean = false) : CPFlagProvider()
 
-abstract class CustomFlagImageProvider : FlagProvider() {
+abstract class CustomCPFlagImageProvider : CPFlagProvider() {
     abstract fun getFlag(alpha2Code: String): Int
 }
