+++ countrypicker/src/main/java/com/hbb20/countrypicker/models/CPCountry.kt
@@ -1,4 +1,4 @@
-package com.hbb20
+package com.hbb20.countrypicker.models
 
 import java.text.Collator
