+++ countrypicker/src/main/java/com/hbb20/countrypicker/models/CPLanguage.kt
@@ -1,4 +1,4 @@
-package com.hbb20
+package com.hbb20.countrypicker.models
 
 import java.util.*
