+++ countrypicker/src/test/java/com/hbb20/CPDataStoreGeneratorTest.kt
@@ -1,6 +1,8 @@
 package com.hbb20
 
 import android.content.res.Resources
+import com.hbb20.countrypicker.datagenerator.CPDataStoreGenerator
+import com.hbb20.countrypicker.datagenerator.CountryFileReading
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
