+++ countrypicker/src/test/java/com/hbb20/CPRecyclerViewHelperTest.kt
@@ -1,8 +1,10 @@
 package com.hbb20
 
 import android.content.res.Resources
-import com.hbb20.countrypicker.CPRecyclerViewHelper
 import com.hbb20.countrypicker.config.CPCountryRowConfig
+import com.hbb20.countrypicker.config.CPRecyclerViewConfig
+import com.hbb20.countrypicker.datagenerator.CPDataStoreGenerator
+import com.hbb20.countrypicker.recyclerview.CPRecyclerViewHelper
 import com.nhaarman.mockitokotlin2.mock
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
@@ -19,7 +21,7 @@ class CPRecyclerViewHelperTest {
 
         val helper = CPRecyclerViewHelper(
             cpDataStore = dataStore,
-            preferredCountryCodes = "IN,AU",
+            cpRecyclerViewConfig = CPRecyclerViewConfig(preferredCountryCodes = "IN,AU"),
             onCountryClickListener = mock()
         )
         assertEquals("IN", helper.allPreferredCountries[0].alpha2)
@@ -30,7 +32,7 @@ class CPRecyclerViewHelperTest {
     fun trimCodeBeforePreferred() {
         val helper = CPRecyclerViewHelper(
             cpDataStore = dataStore,
-            preferredCountryCodes = "IN ,AU",
+            cpRecyclerViewConfig = CPRecyclerViewConfig(preferredCountryCodes = "IN ,AU"),
             onCountryClickListener = mock()
         )
         assertEquals("IN", helper.allPreferredCountries[0].alpha2)
@@ -41,7 +43,7 @@ class CPRecyclerViewHelperTest {
     fun preferredCountriesWorkWithAlpha3() {
         val helper = CPRecyclerViewHelper(
             cpDataStore = dataStore,
-            preferredCountryCodes = "IND,AUS",
+            cpRecyclerViewConfig = CPRecyclerViewConfig(preferredCountryCodes = "IND,AUS"),
             onCountryClickListener = mock()
         )
         assertEquals("IN", helper.allPreferredCountries[0].alpha2)
@@ -52,7 +54,7 @@ class CPRecyclerViewHelperTest {
     fun preferredCountriesWorkWithAlpha2and3() {
         val helper = CPRecyclerViewHelper(
             cpDataStore = dataStore,
-            preferredCountryCodes = "IN,AUS",
+            cpRecyclerViewConfig = CPRecyclerViewConfig(preferredCountryCodes = "IN,AUS"),
             onCountryClickListener = mock()
         )
         assertEquals("IN", helper.allPreferredCountries[0].alpha2)
@@ -63,7 +65,7 @@ class CPRecyclerViewHelperTest {
     fun invalidCodesAvoidedSafely() {
         val helper = CPRecyclerViewHelper(
             cpDataStore = dataStore,
-            preferredCountryCodes = "XX,IN,YYY,21dksaj,AUS",
+            cpRecyclerViewConfig = CPRecyclerViewConfig(preferredCountryCodes = "XX,IN,YYY,21dksaj,AUS"),
             onCountryClickListener = mock()
         )
         assertEquals("IN", helper.allPreferredCountries[0].alpha2)
@@ -74,7 +76,7 @@ class CPRecyclerViewHelperTest {
     fun duplicateCodesAreRemoved() {
         val helper = CPRecyclerViewHelper(
             cpDataStore = dataStore,
-            preferredCountryCodes = "IN,AUS,IN",
+            cpRecyclerViewConfig = CPRecyclerViewConfig(preferredCountryCodes = "IN,AUS,IN"),
             onCountryClickListener = mock()
         )
 
@@ -87,7 +89,7 @@ class CPRecyclerViewHelperTest {
     fun duplicateCountriesAreRemoved() {
         val helper = CPRecyclerViewHelper(
             cpDataStore = dataStore,
-            preferredCountryCodes = "IN,AUS,IND",
+            cpRecyclerViewConfig = CPRecyclerViewConfig(preferredCountryCodes = "IN,AUS,IND"),
             onCountryClickListener = mock()
         )
